Fixes #88
Adds a new transition state for replicators: Stopping.
This ends in the Stopped state, which can be fired
multiple times (such once by UpdateActive, and another
manually by unit test).

Updates the unit test TearDown to take advantage of
this event to let activity drain before closing
the database and manager. Previously, this was
occuring to early, and as a result certain
tasks were failing as the database reference
was cleared out already.

This commit also adds an IsOpen check to the storage-
engine's Open method, avoiding attempts to open an
already open database, which yields an out of memory
error by sqlite.